added a new ImageCarousel that combines antd's Carousel with photoswipe (so you can click to zoom in). Added the navbar to the Login page
