+++ packages/frontend/app/lti/(auth)/login/page.tsx
@@ -1,5 +1,12 @@
 'use client'
 
-import LoginPage from '@/app/(auth)/login/page'
+import LoginPage from '@/app/(auth)/login/components/LoginPage'
 
-export default LoginPage
+/* Same as regular login page, just without the navbar header */
+export default function LoginLTI() {
+  return (
+    <>
+      <LoginPage />
+    </>
+  )
+}
