+++ packages/frontend/package.json
@@ -9,10 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@ant-design/cssinjs": "^1.24.0",
     "@ant-design/icons": "^5.3.7",
     "@ant-design/nextjs-registry": "^1.1.0",
     "@ant-design/v5-patch-for-react-19": "^1.0.3",
-    "@ant-design/cssinjs": "^1.24.0",
     "@fullcalendar/common": "^5.11.5",
     "@fullcalendar/react": "^6.1.15",
     "@radix-ui/react-dialog": "^1.1.1",
@@ -37,8 +37,10 @@
     "json-to-csv-export": "^2.1.1",
     "katex": "^0.16.22",
     "lucide-react": "^0.503.0",
-    "next": "^15.3.6",
     "nanoid": "^5.1.5",
+    "next": "^15.3.6",
+    "photoswipe": "^5.4.4",
+    "photoswipe-dynamic-caption-plugin": "^1.2.7",
     "platform": "^1.3.6",
     "prismjs": "^1.30.0",
     "react": "19.1.2",
@@ -64,6 +66,7 @@
   },
   "devDependencies": {
     "@types/node": "^22.15.2",
+    "@types/photoswipe": "^5.2.5",
     "@types/react": "19.1.2",
     "@types/react-dom": "19.1.2",
     "@types/react-google-recaptcha": "^2.1.9",
